Вынес `/help` в отдельный класс
